Automatically add padding to fingerprints
Pad the fingerprints with 0s when the number of fingerprints
isn't evenly divisible by the number of dimensions. An attempt
is made to pad the dimensions evenly.

One TODO is that the codeword array that is written out still
has the padding.

Also fixed a bug where the coordinates weren't removed from the
fingerprints before clustering.