Add version tag to documentation:
The command `git describe --tags` finds the most recent tag and returns
also the number of commits on top of the found tag and the commit hash
of the current commit. This links the renderd documententaion with the
codebase it actually reflects.

If there are no tags, or, if we are not in a git repository,
`result.stdout.strip()` is an empty string.