Remove conditional execution of push image step:
The image now gets pushed on PRs to `main` and `develop` and when a git
tag gets pushed. The conditional execution on _approved_ PRs is removed.

If a git tag is pushed (usually on `main` via bumpversion for new
releases), then an associated Docker image tagged with the git tag will
be pushed to the registry.

On PRs a Docker image with tag `latest` will be pushed to the registry.