Enable layer caching on Docker image build:
Moving `ARG` down a few lines in the `Dockerfile` enables caching of the
layers while building the image. For the layers above the `ARG` line
this makes sense to me, why the caching now also works for the layers
below `ARG` is still not clear to me. But it works!