Avoids pipeline trigger on approved reviews:
If the review is approved, a run of the pipeline was triggered but
failed due to insufficient rights from the forked repository. There is
actually no need for an extra run of the pipeline once a review is
approved since a push to master or develop is likely to follow. This
merge/push is done with sufficient rights of the repositories owners.

Fixes #212