Avoid necessity to rename top level directory:
The top level directory in the zip file is suffixed with a commit hash
that relates to the downloaded test data repository. Subsequent steps in
the pipeline expect this directory have the name `test_data`. This snippet
avoids manual renaming of the extracted folder with each newer version
of the test data repository.